cloud_functions/requirements.txt
google-cloud-storage
google-cloud-bigquery
functions-framework
pyarrow
pandas